requirements.txt

flask==2.3.2
google-genai==1.38.0
tiktoken==0.9.0

